+++ plugins/numbers/src/tagger.js
@@ -2,7 +2,7 @@ const multiples =
   '(hundred|thousand|million|billion|trillion|quadrillion|quintillion|sextillion|septillion)'
 
 // improved tagging for numbers
-const tagger = function(doc) {
+const tagger = function (doc) {
   doc.match(multiples).tag('#Multiple')
   //  in the 400s
   doc.match('the [/[0-9]+s$/]').tag('#Plural')
@@ -12,5 +12,12 @@ const tagger = function(doc) {
   doc.match('#Value and a (half|quarter)').tag('Value', 'value-and-a-half')
   //one hundred and seven dollars
   doc.match('#Money and #Money #Currency?').tag('Money', 'money-and-money')
+  // doc.debug()
+  // $5.032 is invalid money
+  doc
+    .match('#Money')
+    .not('#TextValue')
+    .match('/\\.[0-9]{3}$/')
+    .unTag('#Money', 'three-decimal money')
 }
 module.exports = tagger