+++ src/02-tagger/04-correction/matches/08-place.js
@@ -27,6 +27,13 @@ let list = [
     tag: 'Region',
     reason: 'district-of-Foo',
   },
+  // in Foo California
+  {
+    match: 'in [#ProperNoun] #Place',
+    group: 0,
+    tag: 'Place',
+    reason: 'propernoun-place',
+  },
 
   // ==== Address ====
   {