+++ changelog.md
@@ -8,6 +8,26 @@ compromise uses semver, and pushes to npm frequently
 
 While all _Major_ releases should be reviewed, our only two _large_ releases are **v6** in 2016 and and **v12** in 2019. Others have been mostly incremental, or niche.
 
+#### [Unreleased]
+
+-move (some) contractions into tokenizer?
+
+#### 13.2.0
+
+- deprecate `.money()` and favour overloaded method in compromise-numbers plugin
+- add `.percentages()` and `.fractions()` to compromise-numbers plugin
+- add `.hasAfter()` and `.hasBefore()` methods
+- change handling of slashes
+- add `.world()` method to constructor
+- add more abbreviations
+- fix regex backtracking #739
+- tokenize build:
+- - remove conjugation and inflection data
+- - remove conjugation and inflection functions
+- remove sourcemap from build process (too big)
+- improvements to `.numbers().units()`
+- fix for linked-list runtime error #744 with contractions
+
 #### 13.1.1
 
 - fix `verbs.json()` runtime-error